Extract UpgradeBar into a separate file
